backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
supabase==2.3.5
psycopg2-binary==2.9.9
cryptography==41.0.7
PyJWT>=2.8.1
djangorestframework-simplejwt==5.4.0
django-ratelimit==4.1.0
python-decouple==3.8
requests==2.31.0
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
openpyxl==3.1.0
reportlab==4.0.7

